[fix] 3~5단계 프론트 구현 요구에 맞는 대규모 수정

+++ app/schemas/common.py
@@ -9,16 +9,16 @@ T = TypeVar("T")
 
 class Region(BaseModel):
     page: int
-    x: int
-    y: int
-    w: int
-    h: int
+    x: float
+    y: float
+    w: float
+    h: float
 
 
 class Point(BaseModel):
     page: int
-    x: int
-    y: int
+    x: float
+    y: float
 
 
 # ── 페이지네이션 메타 ───────────────────────────────────────────
